agentic-chat/requirements.txt
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
openai-agents>=0.0.19
